requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bidict==0.22.1
blinker==1.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
DateTime==4.7
distlib==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.0.4
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
path==16.5.0
path.py==12.5.0
platformdirs==3.5.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2022.6
requests==2.31.0
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
trio==0.22.1
trio-websocket==0.10.3
urllib3==1.26.12
virtualenv==20.23.0
Werkzeug==2.3.6
wsproto==1.2.0
WTForms==3.0.1
zope.interface==6.0
